提交表和代码

+++ ExcelToLua/src/ExcelToLua/ExcelToLua/JsonConvertor/JsonTable.cs
@@ -161,7 +161,7 @@ namespace ExcelToLua
             if (m_is_precision)
                 sb.Append(Math.Round(m_data, m_precision));
             else
-                sb.Append(m_data.ToString());
+                sb.Append(((Decimal)m_data).ToString());
         }
     }
